agents/template_agents/requirements_graphrag.txt
# GraphRAG 依赖包
# pip install https://github.com/awslabs/graphrag-toolkit/archive/refs/tags/v3.11.1.zip#subdirectory=lexical-graph
llama-index>=0.10.0
llama-index-core>=0.10.0
boto3>=1.28.0
botocore>=1.31.0
opensearch-py
llama-index-vector-stores-opensearch
python-dotenv

# 数据处理
pyyaml>=6.0
nest-asyncio>=1.5.0
